# docs/swagger.yaml
basePath: /
definitions:
...
  dto.PositionCreateRequest:
    properties:
      attributes:
        type: string
      course:
        type: integer
      device_id:
        type: string
      device_time:
        type: string
      latitude:
        maximum: 90
        minimum: -90
        type: number
      longitude:
        maximum: 180
        minimum: -180
        type: number
      speed:
        type: integer
    required:
    - device_id
    - device_time
    - latitude
    - longitude
    type: object
...
  utils.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta: {}
      status:
        type: boolean
    type: object
host: localhost:8888
info:
  contact:
    email: mddx56@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  description: API para gestionar motos
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: RutaSegura API
  version: 0.1.2
paths:
...
  /api/positions:
    post:
      consumes:
      - application/json
      description: Create a new position record for a device
      parameters:
      - description: Position Create Request
        in: body
        name: position
        required: true
        schema:
          $ref: '#/definitions/dto.PositionCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create a new position
      tags:
      - positions
  /api/positions/{id}:
    get:
      consumes:
      - application/json
      description: Get a position by its ID
      parameters:
      - description: Position ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get position by ID
      tags:
      - positions
  /api/positions/coordinates:
    get:
      consumes:
      - application/json
      description: Get position coordinates (latitude, longitude) for a specific device
        and date for frontend mapping
      parameters:
      - description: IMEI
        in: query
        name: imei
        required: true
        type: string
      - description: 'Date (YYYY-MM-DD format, default: today)'
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get position coordinates by IMEI and date
      tags:
      - positions
  /api/positions/device:
    get:
      consumes:
      - application/json
      description: Get all positions for a specific device defined by IMEI
      parameters:
      - description: IMEI
        in: query
        name: imei
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get positions by IMEI
      tags:
      - positions
  /api/positions/device-details:
    get:
      consumes:
      - application/json
      description: Get positions and vehicle details for admin
      parameters:
      - description: IMEI
        in: query
        name: imei
        required: true
        type: string
      - description: Date (YYYY-MM-DD)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get positions with vehicle info by IMEI and date
      tags:
      - positions
  /api/positions/history:
    get:
      consumes:
      - application/json
      description: Get a detailed history timeline for a device and date (trips, stops,
        gaps)
      parameters:
      - description: Device IMEI
        in: query
        name: imei
        required: true
        type: string
      - description: Date (YYYY-MM-DD)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get device history
      tags:
      - positions
  /api/positions/last:
    get:
      consumes:
      - application/json
      description: Get the last position for a specific device defined by IMEI
      parameters:
      - description: IMEI
        in: query
        name: imei
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get last position by IMEI
      tags:
      - positions
  /api/positions/latest:
    get:
      consumes:
      - application/json
      description: Get the last position for all devices
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get last position of all devices
      tags:
      - positions
  /api/positions/route:
    get:
      consumes:
      - application/json
      description: Get a detailed history timeline for a device within a specific
        time range
      parameters:
      - description: Device IMEI
        in: query
        name: imei
        required: true
        type: string
      - description: Start Time (YYYY-MM-DD HH:mm:ss)
        in: query
        name: start_time
        required: true
        type: string
      - description: End Time (YYYY-MM-DD HH:mm:ss)
        in: query
        name: end_time
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get device route history (Replay)
      tags:
      - positions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"